test: introduce createTestBotContext
this commit introduces a new utility to provision a test bot context.
this makes the testing strategy becomes much simpler, as you already
have a whole context setup for you to call the commands you might wanna
test.

this comes with some goodies: a test user, a prepared i18n instance in
english, and a bag with mocks that makes it more idiomatic to write some
tests.

I suggest using the inline snapshot feature as you have a good
experience seeing the input and output of commands.