tests: improve unit test foundation (#23)
* test: introduce createTestBotContext

this commit introduces a new utility to provision a test bot context.
this makes the testing strategy becomes much simpler, as you already
have a whole context setup for you to call the commands you might wanna
test.

this comes with some goodies: a test user, a prepared i18n instance in
english, and a bag with mocks that makes it more idiomatic to write some
tests.

I suggest using the inline snapshot feature as you have a good
experience seeing the input and output of commands.

* test: add anonymized database and extra tests

still gotta check if mutations done to the anonymized database are persisted in disk. I wouldn't want that. but I might have to create some knobs in the createDatabaseInstance to make it readonly/memory only on test environments

* chore: upgrade to node 19, ESM modules, telegraf, pino, lowdb,
everything

* fix: migrate to vitest

after much struggle with jest on ESM, vitest worked marvelously out of
the box and extremely fast.

* test: add coverage tooling, fix semantics

* test: add more tests

* ci: fix test command

* test: deterministic tests for pingMembersAt

* fix: remove unused import